Creating Stable Ros Launch & Exit #1058
The way ros is currently launched leads to ros not shutting down properly when a node is shut down.

Adding required = true to all ros nodes mean that they can be cleanly exited when one of the nodes is shut down. This ensures a stable setup to use on a repeated basis for trial runs using the SubT simulator.